// apps/frontend/src/api/generated/analyseJobInProgressDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Job Research AI
 * Find job faster with Job Research AI
 * OpenAPI spec version: 1.0.0
 */

export interface AnalyseJobInProgressDto {
  /** eye1243.. */
  id: string;
  /** WAITING */
  status: string;
  /** Junkie 1.0 */
  presetName: string;
}