// apps/frontend/src/api/generated/simplifiedAnalyseJobWithPresetResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Job Research AI
 * Find job faster with Job Research AI
 * OpenAPI spec version: 1.0.0
 */
import type { AnalyseJobInProgressDto } from './analyseJobInProgressDto';

export interface SimplifiedAnalyseJobWithPresetResponse {
  jobs: AnalyseJobInProgressDto[];
}